16.04 fixes

+++ camodocal.sh
@@ -11,11 +11,16 @@ sudo apt install -y libgoogle-glog-dev
 
 echo "HACK TO GET AROUND CMAKE BUILD ERROR IN UBUNTU 16.04:"
 echo "sudo ln -s /usr/lib/x86_64-linux-gnu/libglog.so /usr/local/lib/"
-sudo ln -s /usr/lib/x86_64-linux-gnu/libglog.so /usr/local/lib/
+if [ ! -f /usr/local/lib/libglog.so ]
+then
+    sudo ln -s /usr/lib/x86_64-linux-gnu/libglog.so /usr/local/lib/
+fi
+
+cd ~/src
 
 if [ ! -d ~/src/camodocal ]
 then
-    git clone git@github.com:hengli/camodocal.git
+    git clone https://github.com/hengli/camodocal.git
 fi
 
 cd ~/src/camodocal
